Move platform-specific question to another page, update page validation
